refactor: moved shell constants in a external file
