+++ test/mocks/carbonio-shell-ui.tsx
@@ -11,39 +11,9 @@ import { generateRoots } from './folders/roots-generator';
 import { getSoapFetch } from './network/fetch';
 import { generateSettings } from './settings/settings-generator';
 
+export { FOLDERS, ZIMBRA_STANDARD_COLORS } from './carbonio-shell-ui-constants';
+
 const FakeIntegration = (): JSX.Element => <div data-testid="fake-component" />;
-
-export const FOLDERS = {
-	USER_ROOT: '1',
-	INBOX: '2',
-	TRASH: '3',
-	SPAM: '4',
-	SENT: '5',
-	DRAFTS: '6',
-	CONTACTS: '7',
-	TAGS: '8',
-	CONVERSATIONS: '9',
-	CALENDAR: '10',
-	ROOT: '11',
-	NOTEBOOK: '12', // no longer created in new mailboxes since Helix (bug 39647).  old mailboxes may still contain a system folder with id 12
-	AUTO_CONTACTS: '13',
-	IM_LOGS: '14',
-	TASKS: '15',
-	BRIEFCASE: '16'
-};
-
-export const ZIMBRA_STANDARD_COLORS = [
-	{ zValue: 0, hex: '#000000', zLabel: 'black' },
-	{ zValue: 1, hex: '#2b73d2', zLabel: 'blue' },
-	{ zValue: 2, hex: '#2196d3', zLabel: 'cyan' },
-	{ zValue: 3, hex: '#639030', zLabel: 'green' },
-	{ zValue: 4, hex: '#1a75a7', zLabel: 'purple' },
-	{ zValue: 5, hex: '#d74942', zLabel: 'red' },
-	{ zValue: 6, hex: '#ffc107', zLabel: 'yellow' },
-	{ zValue: 7, hex: '#edaeab', zLabel: 'pink' },
-	{ zValue: 8, hex: '#828282', zLabel: 'gray' },
-	{ zValue: 9, hex: '#ba8b00', zLabel: 'orange' }
-];
 
 const mockedAccount = generateAccount();
 const mockedAccounts = [mockedAccount];
